This is the first half of some re-work on the CSL garbage collection. It changes things so that there is no longer need for a dedicated garbage collector mark bit in every Lisp Object. In the old world addresses all had to have the same sign so I could use the most significant bit as this mark bit, while immediate data has to leave one bit spare for it. WIth the change (which uses a tag code for fowarding addresses) the heap on a 32-bit system can apen across the 2G boundary, and when I do the next stage of rearrangement I will have one more bit available in every header word and every unit of immediate data.
I also had a bug in the serialization code whereby symbols whose name had
non-ASCII characters in (ie ones with more extended Unicode) got corrupted
when passed via the FASL mechanism. Specifically they ended up being adjusted
into UTF-8 twice over. This in particular could damage the end of file
marker.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3829 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/cslread.h
@@ -96,10 +96,10 @@ typedef struct Package
 #define NO_PREFIX           'x'
 
 extern bool is_constituent(int c);
-extern LispObject intern(int len, bool escaped);
-extern LispObject iintern(LispObject str, int32_t h, LispObject p,
+extern LispObject intern(size_t len, bool escaped);
+extern LispObject iintern(LispObject str, size_t h, LispObject p,
                           int str_is_ok /* NOT a bool */);
-extern LispObject find_package(char *name, int len);
+extern LispObject find_package(char *name, size_t len);
 extern LispObject read_from_vector(char *v);
 
 #define MAX_PROMPT_LENGTH 80
